feat(tickets): add tickets management (#7)

+++ src/lib/components/Table.svelte
@@ -3,6 +3,7 @@
 
 	export let headers: string[];
 	export let data: any[] | null;
+	export let preview: boolean = false;
 
 	let baseUrl: string = '';
 
@@ -11,10 +12,19 @@
 	onMount(() => {
 		baseUrl = window.location.pathname;
 	});
+
+	const format = (col: any): string | number => {
+		if (typeof col === 'number' && col.toString().split('.')[1]?.length > 2) {
+			return col.toFixed(2);
+		}
+		return col;
+	};
 </script>
 
 <div class="w-full text-right flex justify-center items-center flex-col">
-	<a class="btn ml-auto bg-green-600" href="{baseUrl}/create"> New</a>
+	{#if !preview}
+		<a class="btn ml-auto bg-green-600" href="{baseUrl}/create"> New</a>
+	{/if}
 	<table class="w-full text-sm text-left mt-6">
 		<thead class="text-dark text-xl">
 			<tr class="bg-gray-300 rounded">
@@ -26,16 +36,22 @@
 		</thead>
 		<tbody class="text-lg my-6">
 			{#if body.length === 0}
-				<tr><td colspan={headers.length}>The data is empty</td></tr>
+				<tr><td class="text-center pt-8" colspan={headers.length}>The data is empty</td></tr>
 			{/if}
 			{#each body as row}
 				<tr class="odd:bg-gray-50 even:bg-gray-100">
 					{#each row as col}
-						<td class="p-4">{col}</td>
+						<td class="p-4">
+							{format(col)}
+						</td>
 					{/each}
 					<td class="text-right">
-						<a class="btn bg-secondary" href="{baseUrl}/edit/{row[0]}"> Edit</a>
-						<a class="btn bg-alternative" href="{baseUrl}/delete/{row[0]}"> Delete</a>
+						{#if !preview}
+							<a class="btn bg-secondary" href="{baseUrl}/edit/{row[0]}"> Edit</a>
+							<a class="btn bg-alternative" href="{baseUrl}/delete/{row[0]}"> Delete</a>
+						{:else}
+							<a class="btn bg-secondary" href="{baseUrl}/see/{row[0]}"> See</a>
+						{/if}
 					</td>
 				</tr>
 			{/each}
